requirements.txt
annotated-types==0.7.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
google-auth==2.38.0
google-genai==1.2.0
idna==3.10
numpy==2.2.3
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websockets==14.2
